+++ cartridges/int_klarna_payments/cartridge/scripts/VCN/klarnaPaymentsVCNSettlement.js
@@ -111,6 +111,10 @@ function createVCNSettlement( order, klarnaPaymentsOrderID, localeObject ) {
         } );
     } catch ( e ) {
         log.error( 'Error in creating Klarna Payments VCN Settlement: {0}', e );
+
+        var KlarnaAdditionalLogging = require( '*/cartridge/scripts/util/klarnaAdditionalLogging' );
+        KlarnaAdditionalLogging.writeLog( order, order.custom.kpSessionId, 'klarnaPaymentsVCNSerrlement.createVCNSettlement()', 'Error in creating Klarna Payments VCN Settlement. Error:' + JSON.stringify( e ) );
+        
         return false;
     }